Replace AppVeyor with Travis-CI
In order to properly cross-compile appix for all three environments, we have to use an OSX build agent.
Since AppVeyor only supports Windows build agents, with this commit we're switching to Travis-CI.
To do this the .travis.yml descriptor has to be added to the root of the repo, and the build script had to be adjusted.