Replace manual mutex locking with channels
To make sure we're not getting into an incosistent state due to multiple goroutines
modifying the watcherState, we were using a Mutex. This is now changed to use communication
on channels instead, and we are changint the watcherState only on from one single goroutine,
so we don't have a race condition any more.